SQLite library almost completed.

+++ src/main/java/com/artingl/easyrg/misc/Regions/RegionPosition.java
@@ -4,11 +4,11 @@ import org.bukkit.Location;
 
 public class RegionPosition {
 
-    private final int x;
-    private final int y;
-    private final int z;
+    private final double x;
+    private final double y;
+    private final double z;
 
-    public RegionPosition(int x, int y, int z) {
+    public RegionPosition(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
@@ -20,15 +20,15 @@ public class RegionPosition {
         this.z = location.getBlockZ();
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
